% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.plrd}
\alias{coef.plrd}
\title{Extract plrd coefficient}
\usage{
\method{coef}{plrd}(object, ...)
}
\arguments{
\item{object}{plrd object}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Extract plrd coefficient
}
