% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.Lipschitz.constant}
\alias{get.Lipschitz.constant}
\title{Function estimating the Lipschitz constant for the conditional response functions}
\usage{
get.Lipschitz.constant(y, x, threshold, diff.curvatures = F, alpha.B = 0.05)
}
\arguments{
\item{y}{The outcomes.}

\item{x}{The running variable.}

\item{threshold}{The threshold for treatment.}

\item{diff.curvatures}{Whether we consider different curvatures before and after the threshold.}

\item{alpha.B}{Significance threshold we use in estimation of the Lipschitz constant.}
}
\value{
Lipschitz constant for the conditional response functions.
}
\description{
Function estimating the Lipschitz constant for the conditional response functions
}
\keyword{internal}
