% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plrd.R
\name{plrd.optim}
\alias{plrd.optim}
\title{plrd.optim: The optimizer we use to implement plrd with cross-fitting}
\usage{
plrd.optim(
  Y = NULL,
  X = NULL,
  threshold = NULL,
  W = NULL,
  Lipschitz.constant = NULL,
  sigma.sq = NULL,
  diff.curvatures = FALSE,
  alpha = NULL,
  bin.width = NULL,
  num.bucket = NULL,
  use.spline = TRUE,
  spline.df = NULL,
  use.homoskedastic.variance = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Y}{The outcomes.}

\item{X}{The running variable.}

\item{threshold}{Sharp Regression Discontinuity threshold "c" at which CATE is to be estimated.}

\item{W}{The treatment indicators, default is set as NULL for sharp RDD.}

\item{Lipschitz.constant}{A Lipschitz bound for the second derivative of mu_w(x) = E(Y(w) | X = x).}

\item{sigma.sq}{An estimate of the homoskedastic variance of Y conditional on X}

\item{diff.curvatures}{Set true if user believes the curvatures are different before and after the threshold c.}

\item{alpha}{Coverage probability of confidence intervals.}

\item{bin.width}{Bin width for discrete approximation.}

\item{num.bucket}{Number of bins for discrete approximation. Can only be used if bin.width = NULL.}

\item{use.spline}{Whether non-parametric components should be modeled as quadratic splines
in order to reduce the number of optimization parameters, and potentially
improving computational performance.}

\item{spline.df}{Number of degrees of freedom (per running variable) used for spline computation.}

\item{use.homoskedastic.variance}{Whether confidence intervals should be built assuming homoskedasticity.}

\item{verbose}{whether the optimizer should print progress information}
}
\value{
Parts of the desired plrd object for the specific sample split
}
\description{
plrd.optim: The optimizer we use to implement plrd with cross-fitting
}
