% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plrd.R
\name{plrd}
\alias{plrd}
\title{plrd: Partially Linear Regression Discontinuity Inference}
\usage{
plrd(
  Y,
  X,
  threshold,
  W = NULL,
  max.window = NULL,
  Lipschitz.constant = NULL,
  sigma.sq = NULL,
  alpha = 0.95,
  diff.curvatures = FALSE,
  signif.curvature = 0.001,
  bin.width = NULL,
  num.bucket = 400,
  use.spline = TRUE,
  spline.df = 40,
  seed = 42,
  use.homoskedastic.variance = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Y}{The outcomes.}

\item{X}{The running variable.}

\item{threshold}{Sharp Regression Discontinuity threshold "c" at which CATE is to be estimated.}

\item{W}{The treatment indicator. Currently plrd only supports sharp RDD, hence W is set as NULL by default.}

\item{max.window}{A reasonable window for estimating the Lipschitz bound on curvature accurately. Default is the whole range.}

\item{Lipschitz.constant}{A Lipschitz bound for the second derivative of mu_w(x) = E(Y(w) | X = x).}

\item{sigma.sq}{An estimate of the homoskedastic variance of Y conditional on X}

\item{alpha}{Coverage probability of confidence intervals.}

\item{diff.curvatures}{Set true if user believes the curvatures are different before and after the threshold c.}

\item{signif.curvature}{Significance level for testing whether curvature is different before and after the threshold.}

\item{bin.width}{Bin width for discrete approximation.}

\item{num.bucket}{Number of bins for discrete approximation. Can only be used if bin.width = NULL.}

\item{use.spline}{Whether non-parametric components should be modeled as cubic splines
in order to reduce the number of optimization parameters, and potentially
improving computational performance.}

\item{spline.df}{Number of degrees of freedom (per running variable) used for spline computation.}

\item{seed}{Random seed for reproducibility.}

\item{use.homoskedastic.variance}{Whether confidence intervals should be built assuming homoskedasticity. As default we use FALSE (i.e., use heteroscedastic standard errors).}

\item{verbose}{Whether the optimizer should print progress information.}
}
\value{
A trained plrd object containing the PLRD estimator, half width of confidence interval and other auxiliary objects such as MSE-optimal weights, worst-case bias, Lipschitz constant, etc.
}
\description{
Optimized estimation and bias-aware inference for treatment effects identified by regression discontinuities
under partially linear model as proposed by Ghosh, Imbens and Wager (2025).
}
\examples{
\donttest{
# Simple example of regression discontinuity design
set.seed(42)
n = 1000; threshold = 0
X = runif(n, -1, 1)
W = as.numeric(X >= threshold)
Y = (1 + 2*W)*(1 + X^2) + 1 / (1 + exp(X)) + rnorm(n, sd = .5)
out = plrd(Y, X, threshold)
print(out)
plot(out)
}

}
\references{
Ghosh, A., Imbens, G., & Wager, S. (2025).
PLRD: Partially Linear Regression Discontinuity Inference.
arXiv preprint arXiv:2503.09907.
}
