% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.plrd}
\alias{summary.plrd}
\title{plrd summary}
\usage{
\method{summary}{plrd}(object, ...)
}
\arguments{
\item{object}{plrd object}

\item{...}{Additional arguments (currently ignored).}
}
\description{
plrd summary
}
