% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.plrd}
\alias{coef.plrd}
\title{Extract plrd coefficient}
\usage{
\method{coef}{plrd}(object, ...)
}
\arguments{
\item{...}{Additional arguments (currently ignored).}

\item{x}{plrd object}
}
\description{
Extract plrd coefficient
}
