% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.plrd}
\alias{plot.plrd}
\title{Plot a plrd object}
\usage{
\method{plot}{plrd}(
  x,
  type = "default",
  percentage.cumulative.weights = 0.99,
  spline.df = 3,
  ...
)
}
\arguments{
\item{x}{plrd object}

\item{type}{Type of plot the user wants to see. We offer three options: "default", "weights", and "combined". The "default" option shows the scatterplot of the original data with black curves that are representative regression functions in our data-driven function class. The dashed line shows the threshold, and the dotted lines indicate the window containing a percentage (99\% by default) of the cumulative absolute plrd weights. The "weights" option plots plrd weights (note, two sets of plrd weights because we use cross-fitting). The "combined" option is a fancy plot combining both the scatterplot and the plot of plrd weights.}

\item{percentage.cumulative.weights}{The percentage of the cumulative absolute weights user wants to keep (for visualization purposes only)}

\item{spline.df}{Degrees of freedom of the natural spline used to plot a representative member of the data-driven function class, constrained to satisfy the smoothness condition.}

\item{...}{Additional graphical arguments to customize the plot, such as \code{xlim}, \code{ylab}, \code{main}, etc.}
}
\description{
Plot a plrd object
}
