% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.plrd}
\alias{plot.plrd}
\title{Plot a plrd object}
\usage{
\method{plot}{plrd}(
  x,
  type = "default",
  percentage.cumulative.weights = 0.99,
  spline.df = 3,
  ...
)
}
\arguments{
\item{x}{plrd object}

\item{type}{The type of plot to display. We offer three options: "default", "weights", and "combined".
The "default" option overlays the scatterplot of the original data with black curves showing representative regression functions in our data-driven function class.
The "weights" option displays the PLRD weights. Shows two sets of PLRD weights because we use cross-fitting with two folds.
The "combined" option displays both the scatterplot and the plot of PLRD weights.
The dashed line marks the threshold, and the dotted lines indicate a post-hoc effective support window containing a percentage of the cumulative absolute PLRD weights.}

\item{percentage.cumulative.weights}{Percentage of cumulative absolute PLRD weights to include in the visualization (determines the post-hoc effective support).}

\item{spline.df}{Degrees of freedom of the natural spline used to plot a representative member of the data-driven function class, constrained to satisfy the smoothness condition of PLRD.}

\item{...}{Additional graphical arguments to customize the plot, such as \code{xlim}, \code{ylab}, \code{main}, etc.}
}
\value{
A list of plot coordinates for the main plot and the weight plot, including the upper and lower bound of the window that contains the specified percentage of cumulative weights.
}
\description{
Plot a plrd object
}
