% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.plrd}
\alias{plot.plrd}
\title{Plot a plrd object}
\usage{
\method{plot}{plrd}(
  x,
  type = "default",
  percentage.cumulative.weights = 0.99,
  spline.df = 3,
  ...
)
}
\arguments{
\item{x}{A fitted \code{plrd} object returned by \code{plrd()}.}

\item{type}{The type of plot to display: \code{"default"}, \code{"weights"},
or \code{"combined"}.}

\item{percentage.cumulative.weights}{Percentage of cumulative absolute PLRD
weight mass used to determine the displayed window on each side of the
threshold. The default is 0.99, providing a visual approximation to the
estimator's effective support.}

\item{spline.df}{Degrees of freedom of the natural spline used to plot a
representative member of the data-driven function class, constrained to
satisfy the smoothness condition of PLRD.}

\item{...}{Additional graphical arguments to customize the plot, such as
\code{xlim}, \code{ylab}, \code{main}, etc.}
}
\value{
A list of plot coordinates for the main plot and the weight plot,
including the upper and lower bounds of the window containing the specified
percentage of cumulative absolute weights.
}
\description{
We offer three plot types: "default", "weights", and "combined". These plots
are constructed post hoc to visualize different features of the fitted PLRD
procedure.
}
\details{
The "default" option overlays the scatterplot of the original data with a
pair of black curves showing representative regression functions in our
data-driven function class. These curves are fitted for visualization and
are not used to compute the PLRD estimate.

The "weights" option displays the fitted PLRD weights. It shows two sets of
weights because we use two-fold cross-fitting.

The "combined" option displays both the default plot and the PLRD weights.

The dashed line marks the threshold, and the dotted lines indicate the window
containing a specified percentage of the cumulative absolute PLRD weight
mass on each side of the threshold.
}
