% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.plrd}
\alias{plot.plrd}
\title{Plot a plrd object}
\usage{
\method{plot}{plrd}(x, percentage.cumulative.weights = 0.99, ...)
}
\arguments{
\item{x}{plrd object}

\item{percentage.cumulative.weights}{The percentage of the cumulative absolute weights user wants to keep (for visualization purposes only)}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Plot a plrd object
}
