% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.plrd}
\alias{print.plrd}
\title{Print a plrd object}
\usage{
\method{print}{plrd}(x, digits = NULL, ...)
}
\arguments{
\item{x}{plrd object}

\item{digits}{number of digits to print}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a plrd object
}
