% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.plrd}
\alias{print.plrd}
\title{Print a plrd object}
\usage{
\method{print}{plrd}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{plrd object}

\item{digits}{number of digits to print}

\item{...}{Additional arguments passed to print methods.}

\item{percentage.cumulative.weights}{The percentage of the cumulative absolute weights to determine effective sample size}
}
\description{
Print a plrd object
}
